Generate theme-settings SCSS file from config YML
