+++ .gitignore
@@ -4,6 +4,7 @@ node_modules
 pattern-lab
 source/_data/design-tokens.artifact.yml
 source/_patterns/00-config/_design-tokens.artifact.scss
+source/_patterns/_uswds-theme-settings.scss
 css/*.css
 css/*.map
 js/dist/*.js
