+++ lib/renderSass.js
@@ -5,28 +5,7 @@
 const { NaniError } = require('nani');
 
 const SassValue = require('./SassValue');
-
-/**
- * Prepares a JS value for output as a Sass value.
- *
- * @param {unknown} value
- */
-function cleanValue(value) {
-  if (typeof value === 'number') {
-    return value;
-  }
-
-  if (typeof value !== 'string') {
-    throw new NaniError({
-      shortMessage: `Value passed to cleanValue() is ${typeof value}, not a string or number`,
-    });
-  }
-
-  if (!value.includes(' ')) {
-    return value;
-  }
-  return value.includes("'") ? '"' + value + '"' : "'" + value + "'";
-}
+const cleanValue = require('./cleanValue');
 
 function createSassMap(data, indent = 2) {
   let output = '';
