+++ source/_patterns/00-config/config.design-tokens.yml
@@ -1,5 +1,9 @@
 # Design Tokens
 # Define basic design variables here.
+uswds:
+  image-path: '../images'
+  root-font-size: 16px
+  font-weight-semibold: 600
 gesso:
   palette:
     brand:
