--- source/_patterns/_uswds-theme-settings.scss
@@ -1,1 +1,0 @@
-$theme-font-weight-semibold: 600;
